Changed code to use specific XPathFactoryImpl
This changes the XPathFactory call to use a specific XPathFactoryImpl to avoid issues where an environment may be using a default implementation that's incompatible with the code (such as using an outdated Saxon parser). This change attempts to use the default Java implementation, but if that is unavailable will fallback to try to load the default.

This will also help to ensure the XML is always parsed in a consistent manor, since everyone should be using the same XPathFactoryImpl.