Clarify the real nature of the RelayState parameter for SSO and SLO
In the most simple case, the RelayState may be used as a "returnUrl",
but I think it's important to underline that the RelayState does not
necessarily need to be a return URL. Indeed, the SAML 2.0 specification
clarifies that a limit of max 80 characters exists for it (at least
in the case of the HTTP-Redirect binding) and that a protection method
against tampering is suggested.
Therefore, a return URL in general would probably be a non-ideal use of
the RelayState parameter, so let's give the latter the relevance it
deserves.