Make SamlResponse more extensible
The query and queryAssertion methods are really useful when creating
extensions of SamlResponse that need to expose other information that
the plain java-saml library does not (like the response issue instant,
or any custom contents of the AuthnContext in the assertion auth
statement). Having these two methods private forces the extender to
reimplement them from scratch, probably by copy-and-paste, which is
clearly sub-optimal.