Polish LogoutRequest constructors and Auth.logout overloadings
This refactor is pretty much the same done with AuthnRequest and
AuthnRequestParams: in this case, the params class encapsulates all the
logout request input parameters used whenever a new logout request is
created for subsequent sending. This as well allows to reduce the number
of LogoutRequest constructor and Auth.logout() overloadings, allows
an extension to use a customized input param object which is also passed
to postProcessXml and eases Auth extensibility.