Provide a setter for validationException
If a subclass needs to extend isValid(String), it must be able to set
the validation exception. In fact, overriding any of the non-private
validateXXX methods is trivial, because a ValidationError can simply be
thrown, but if additional validation is required, which is not covered
by any of those validateXXX methods, a subclass may need to override
isValid(String) directly, and this does not throw ValidationError, but
rather catches it so that it can be subsequently returned by
getValidationException() to the consumer. Hence, a setter with at least
protected visibility is needed for subclasses to do the same.