Split Response and Assertion Issuer retrieval
The SamlResponse.getIssuers() contract is quite controversial. For a
valid response, it will always return just one element. For an invalid
response, depending on the cause it may:
- fail if no Assertion is present: this means, in particular, that if
the status code is not a success one, it's impossible to retrieve the
Response issuer with this method (although it may be a reasonable
requirement, for logging purposes for instance)
- fail if multiple Assertions are present: again, the Response Issuer
cannot be retrieved in this case either
- fail in the unlikely event that multiple Response Issuers were found
- return up to 2 issuers at most, if different issuers were set on the
Response and on the Assertion (which will make isValid() fail), with the
inability to determine which is the Response Issuer and which is the
Assertion Issuer (indeed: the former will be the first one in the
list, the latter will be the second, but this contract is a bit weak)

For these reasons, two different methods were provided to retrieve the
Response and the Assertion Issuers, with the former that will succeed
even when the status code is not a successful one. Also, because of the
above reasons, the getIssuers() method was deprecated in favour of the
two new ones.