Support more complex response statuses in LogoutResponse generation
An overload of build() has been added to supply a SamlResponseStatus,
so that it's possible to specify a more complex response status in the
generated LogoutResponse. A typical example is the case of a partial
logout, which needs a top-level
urn:oasis:names:tc:SAML:2.0:status:Responder status code and a
second-level
urn:oasis:names:tc:SAML:2.0:status:PartialLogout one. Also, a status
message can also be added now.

Closes #342.