Make Name ID and attribute value trimming an opt-in feature
Name IDs (including issuers) are by default left untouched, as well as
attribute values, like it was before. However two new settings have
been introduced (whose default value is false) which allow to enable
trimming for such values, which is probably the desired behaviour in
practice, although SAML specification says that no whitespace processing
should be performed on strings.
Another place where trimming may be desirable is in SessionIndex
extraction from LogoutRequests: this is not performed at any point of
the LogoutRequest processing, but an overloading has been provided so
that the API consumer may still request trimming.
AuthnResponseTest.testGetIssuersTrimming() is disabled by now because it
fails due to a bug in SamlResponse.getIssuers() which is addressed by
another PR.