Polish AuthnRequest constructor and Auth.login overloadings
Input parameters, which drive how the AuthnRequest should be generated,
are now encapsulated into an AuthnRequestParams object which is far more
extensible and help to polish the API of AuthnRequest constructors and
Auth.login, which are crowded with overloadings and cannot be easily
extended.
Also, these input params are passed to postProcessXml: in this way, if
an extension plans to use a specialization of AuthnRequestParams, it can
access such a specialized instance also within a proper postProcessXml
overriding.