Revise SP contacts settings validation
The contact type check now becomes useful (and so it was restored),
because with the new full Contacts support the user may indeed specify
invalid contact types in settings.
The "not enough data" check, instead, was fixed and it is now raised
only if ALL of the contact data (company, given name, surname, e-mail
addresses and phone names) are empty, reflecting the actual SAML 2.0
metadata schema constraint.

Fixes #353.